bulk card count methods
